+++ test/sandbox/config/environment.rb
@@ -4,4 +4,4 @@
 require File.expand_path("../application", __FILE__)
 
 # Initialize the rails application
-Dummy::Application.initialize!
+Sandbox::Application.initialize!