+++ test/sandbox/config/environments/test.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-Dummy::Application.configure do
+Sandbox::Application.configure do
   # Settings specified here will take precedence over those in config/application.rb
 
   # The test environment is used exclusively to run your application's
@@ -27,6 +27,7 @@ Dummy::Application.configure do
 
   config.view_component.preview_paths << "#{Rails.root}/lib/component_previews"
   config.view_component.render_monkey_patch_enabled = true
+  config.view_component.show_previews_source = true
   config.view_component.test_controller = "IntegrationExamplesController"
 
   # Tell Action Mailer not to deliver emails to the real world.