+++ test/test_helper.rb
@@ -18,7 +18,7 @@ require "minitest/autorun"
 # Configure Rails Envinronment
 ENV["RAILS_ENV"] = "test"
 
-require File.expand_path("../config/environment.rb", __FILE__)
+require File.expand_path("../sandbox/config/environment.rb", __FILE__)
 require "rails/test_help"
 
 # Sets custom preview paths in tests.
@@ -49,6 +49,14 @@ def with_preview_controller(new_value)
   yield
   Rails.application.config.view_component.preview_controller = old_value
   app.reloader.reload!
+end
+
+def with_custom_component_path(new_value)
+  old_value = ViewComponent::Base.view_component_path
+  ViewComponent::Base.view_component_path = new_value
+  yield
+ensure
+  ViewComponent::Base.view_component_path = old_value
 end
 
 def with_new_cache